Line-up interfaces to use size_t for GetCount()s (and count related api).
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@38030 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/variant.h
@@ -1,5 +1,5 @@
 /////////////////////////////////////////////////////////////////////////////
-// Name:        variant.h
+// Name:        wx/variant.h
 // Purpose:     wxVariant class, container for any type
 // Author:      Julian Smart
 // Modified by:
@@ -207,7 +207,7 @@ public:
     bool IsValueKindOf(const wxClassInfo* type) const;
 
     // Return the number of elements in a list
-    int GetCount() const;
+    size_t GetCount() const;
 
 // Value accessors
     double GetReal() const ;
@@ -245,7 +245,7 @@ public:
     bool Member(const wxVariant& value) const;
 
     // Deletes the nth element of the list
-    bool Delete(int item);
+    bool Delete(size_t item);
 
     // Clear list
     void ClearList();
